// src/Models/agent-cust.model.ts
export class customer {
    custId: number = 0;
    custFirstName: string = '';
    custLastName: string = '';
    custEmail: string = '';
    custPhoneNumber: number = 0;
    state: string = '';
    city: string = '';
    password: string = '';
    userName: string = '';
    userId:number=0;
    status: boolean = true;
}